requirements.txt
#dotenv
#typer
#phidata
duckduckgo-search
fastapi
uvicorn
groq
#sqlalchemy
#pgvector
#psycopg[binary]
pypdf
openai
requests
#sentence-transformers
# requirements.txt
phi
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
pgvector>=0.2.0
typer>=0.9.0